Moved tests into integrationTests folder
